# Codigo/backend/src/models/review.js
'use strict';
const {
  Model
} = require('sequelize');
module.exports = (sequelize, DataTypes) => {
  class Review extends Model {
    /**
     * Helper method for defining associations.
     * This method is not a part of Sequelize lifecycle.
     * The `models/index` file will call this method automatically.
     */
    static associate(models) {
      Review.belongsTo(models.Movie, { foreignKey: 'movie_id' });
    }
  }
  Review.init({
    movie_id: DataTypes.INTEGER,
    comment: DataTypes.TEXT,
    rating: DataTypes.DECIMAL
  }, {
    paranoid: true,
    sequelize,
    modelName: 'Review',
  });
  return Review;
};